move pin test
