+++ cypress/components/pin/Pin.svelte
